Refactoring: renamed namespaces
