+++ cs/TagsCloudVisualization/Core/ArchimedeanSpiral.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Drawing;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualization.Core
 {
     class ArchimedeanSpiral
     {
