+++ cs/TagsCloudVisualization/Core/Axis.cs
@@ -1,4 +1,4 @@
-﻿namespace TagsCloudVisualization
+﻿namespace TagsCloudVisualization.Core
 {
     enum Axis
     {
