+++ cs/TagsCloudVisualization/Core/CircularCloudLayouter.cs
@@ -1,9 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Drawing;
+using System.Linq;
+using TagsCloudVisualization.Extensions;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualization.Core
 {
     class CircularCloudLayouter
     {
