+++ cs/TagsCloudVisualization/Tests/CircularCloudLayouterTests.cs
@@ -6,8 +6,9 @@ using System.Linq;
 using FluentAssertions;
 using NUnit.Framework;
 using NUnit.Framework.Interfaces;
+using TagsCloudVisualization.Core;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualization.Tests
 {
     [TestFixture]
     class CircularCloudLayouterTests
